Remove aggregators
Primarily, this removes aggregators, which have been deprecated since
1.4 and have been replaced by combiners. This removal sets us up to
provide aggregators as a separate jar artifact, outside of the main
Accumulo build, if there is a continued demand for them. In future, all
such "user library" type add-ons to Accumulo could be provided as
separate, optional component jars.

Also cleanup other minor deprecated items.

+++ core/src/main/java/org/apache/accumulo/core/client/impl/ScannerIterator.java
@@ -48,7 +48,7 @@ public class ScannerIterator implements Iterator<Entry<Key,Value>> {
   private static final Logger log = LoggerFactory.getLogger(ScannerIterator.class);
 
   // scanner options
-  private int timeOut;
+  private long timeOut;
 
   // scanner state
   private Iterator<KeyValue> iter;
@@ -107,7 +107,7 @@ public class ScannerIterator implements Iterator<Entry<Key,Value>> {
   }
 
   ScannerIterator(ClientContext context, Table.ID tableId, Authorizations authorizations,
-      Range range, int size, int timeOut, ScannerOptions options, boolean isolated,
+      Range range, int size, long timeOut, ScannerOptions options, boolean isolated,
       long readaheadThreshold) {
     this.timeOut = timeOut;
     this.readaheadThreshold = readaheadThreshold;
